Use inventory for component plugins
This commit adds a new mechanism for registering component plugin
configurations under a static label. It should be possible to outright
replace the existing method without much hassle, and this new way of
doing it allows us to iterate all plugin implementations of a type.

This will allow us in future to add usability improvements such as
component listing, configuration example generation, etc.

Signed-off-by: Ashley Jeffs <ash@jeffail.uk>